+++ src/lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(not(any(test, feature = "std")), no_std)]
-#![doc(html_root_url = "https://docs.rs/geometrid/0.6.0")]
+#![doc(html_root_url = "https://docs.rs/geometrid/0.7.0")]
 // #![deny(missing_docs)]
 #![allow(warnings, dead_code, unused_imports, unused_mut)]
 #![warn(clippy::pedantic)]
@@ -18,6 +18,7 @@ pub mod shape;
 pub mod tile;
 pub mod tile_map;
 pub mod tile_set;
+pub mod tile_by_row_iter;
 #[cfg(any(test, feature = "u256"))]
 pub mod tile_set256;
 pub mod vector;
@@ -38,6 +39,7 @@ pub mod prelude {
     pub use crate::tile::*;
     pub use crate::tile_map::*;
     pub use crate::tile_set::*;
+    pub use crate::tile_by_row_iter::*;
     #[cfg(any(test, feature = "u256"))]
     pub use crate::tile_set256::*;
     pub use crate::vector::*;